prototyping/function_code/requirements.txt
google-cloud-videointelligence
google-cloud-video-transcoder
google-cloud-storage
